updated feature1 tests
