+++ tests/unit/test_get_all_movies.py
@@ -1,42 +1,23 @@
 # TODO: Feature 1
-from flask import render_template_string
 import pytest
-from flask.testing import FlaskClient
-from app import app
 from src.repositories.movie_repository import get_movie_repository
 
-@pytest.fixture(scope='module')
-def test_app():
-    with app.app_context():
-        yield app.test_client()
-
-def test_list_movies(test_app: FlaskClient):
+def test_list_movies():
     #add movie
-    get_movie_repository().create_movie("mov", "dir", 2)
+    movie_repo = get_movie_repository()
+    new_movie = movie_repo.create_movie("mov", "dir", 2)
 
-    #checc if movie ther
-    response = test_app.get('/movies')
-    response_data = response.data.decode('utf-8')
-    assert response.status_code == 200
-    print(response_data)
+    #assert its there
+    assert new_movie.movie_id in movie_repo.get_all_movies()
 
-    #i expect that the movie, the director, and two filled stars will show up followed by an empty star
-    #idk if github will allow all these qhitespaces but i refuse to change
-    #the spacing in the jinja stuff for readability
-    expectedStuf = '''<td>mov</td>
-            <td>dir</td>
-            <td>
-            
-                
-                    <img src="https://pluspng.com/img-png/star-png-star-png-image-2156.png" style="height:15px;">
-                
-            
-                
-                    <img src="https://pluspng.com/img-png/star-png-star-png-image-2156.png" style="height:15px;">
-                
-            
-                
-                    <img src="https://cdn.onlinewebfonts.com/svg/img_548678.png" style="height:17px">'''
-    print("Expected:")
-    print(expectedStuf)
-    assert expectedStuf in response_data
+    #assert its got stuff
+    assert "mov" is new_movie.title
+    assert "dir" is new_movie.director
+    assert 2 is new_movie.rating
+
+    #take it back
+    new_movie = movie_repo.get_all_movies()[new_movie.movie_id]
+    #make sure its still got stuff
+    assert "mov" is new_movie.title
+    assert "dir" is new_movie.director
+    assert 2 is new_movie.rating
